+++ suomen-vaylat-app/src/components/published-map/PublishedMap.jsx
@@ -222,36 +222,8 @@ const PublishedMap = () => {
             });
 
             // save start state
-            channel.getCurrentState(data => {
-                channel.getSelectedLayers(layers => {
-                    if (data.mapfull && data.mapfull.state) {
-                        const mapfullState = data.mapfull.state;
-                        const selectedLayers = [];
-                        const visibleLayers = [];
-
-                        layers.forEach(layer => {
-                            visibleLayers.push(layer.id);
-                        });
-
-                        // check layer orders
-                        const layerOrders = mapfullState.selectedLayers.filter(layer => {
-                            return visibleLayers.includes(layer.id);
-                        });
-
-                        layerOrders.forEach(layer => {
-                            selectedLayers.push(layer);
-                        });
-
-                        const startState = {
-                            x: mapfullState.east,
-                            y: mapfullState.north,
-                            selectedLayers: selectedLayers,
-                            zoom: mapfullState.zoom
-                        }
-
-                        store.dispatch(setStartState(startState));
-                    }
-                });
+            channel.getPublishedMapState(function (data) {
+                store.dispatch(setStartState(data));
             });
         });