+++ suomen-vaylat-app/src/components/search/VKMTrackSearch.jsx
@@ -2,10 +2,6 @@ import { useState } from 'react';
 import styled from 'styled-components';
 import strings from '../../translations';
 
-import { useAppSelector } from '../../state/hooks';
-
-import { VKMGeoJsonHoverStyles, VKMGeoJsonStyles } from './VKMSearchStyles';
-
 const StyledContainer = styled.div`
     display: flex;
     flex-direction: column;
@@ -26,6 +22,13 @@ const StyledDropdownContentItem = styled.div`
 `;
 
 const StyledDropdownContentItemTitle = styled.p`
+    text-align: ${props => props.type === 'noResults' && 'center'};
+    font-weight: bold;
+    font-size: 14px;
+    color: #504d4d;
+`;
+
+const StyledDropdownContentItemSubtitle = styled.p`
     text-align: ${props => props.type === 'noResults' && 'center'};
     font-size: 14px;
     color: #504d4d;
@@ -53,100 +56,43 @@ const StyledInput = styled.input`
     &:focus {
         outline: none;
     };
-    /*border-left: ${props => props.required ? '2px solid #c73f00' : 'none'};*/
 `;
 
 const StyledLabel = styled.label`
     font-size: 14px;
     font-weight: bold;
     margin: 0px;
+    color: ${props => props.error && props.theme.colors.secondaryColor6};
 `;
 
 const VKMTrackSearch = ({
-  setIsSearching,
   searchValue,
   setSearchValue,
-  setLastSearchValue,
-  vectorLayerId,
-  removeMarkersAndFeatures,
-  setSearchResults
+  handleVKMTrackSearch,
+  vkmTrackError,
 }) => {
-    const [error, setError] = useState(null);
-
 
     const [requiredRatanumero, setRequiredRatanumero] = useState(true);
     const [requiredRatakilometri, setRequiredRatakilometri] = useState(true);
     const [requiredRatametri, setRequiredRatametri] = useState(true);
 
-    const rpc = useAppSelector((state) => state.rpc);
-
-    const handleResponse = (data) => {
-        setIsSearching(false);
-
-        let featureStyle = VKMGeoJsonStyles['track'];
-        let hover = VKMGeoJsonHoverStyles['track'];
-
-        rpc.channel.postRequest('MapModulePlugin.RemoveFeaturesFromMapRequest', [null, null, vectorLayerId + '_vkm_track']);
-
-        const value = {
-            ratanumero: data.hasOwnProperty('ratanumero') ? data.ratanumero : searchValue.ratanumero || '',
-            ratakilometri: data.hasOwnProperty('ratakilometri') ? parseInt(data.ratakilometri) : searchValue.ratakilometri || 1,
-            ratametri: data.hasOwnProperty('ratametri') ? parseInt(data.ratametri) : searchValue.ratakilometri || 0
-        };
-
-        setSearchValue(value);
-        setLastSearchValue(value);
-        setSearchResults(data);
-
-        data.hasOwnProperty('geom') && rpc.channel.postRequest('MapModulePlugin.AddFeaturesToMapRequest',
-        [data.geom, {
-            centerTo: true,
-            hover: hover,
-            featureStyle: featureStyle,
-            layerId: vectorLayerId + '_vkm_track',
-            maxZoomLevel: 10
-        }]);
-    };
-
-    if (searchValue === '' && error !== null) {
-        setError(null);
-    }
-
-    const handleVKMSearch = (params) => {
-        removeMarkersAndFeatures();
-        setIsSearching(true);
-        setError(null);
-
-        if (params.hasOwnProperty('vkmRatanumero') && params.hasOwnProperty('vkmRatakilometri') && params.hasOwnProperty('vkmRatametri')
-            && params.vkmRatanumero !== '' && params.vkmRatakilometri !== '' && params.vkmRatametri !== ''
-        ) {
-            rpc.channel.searchVKMTrack && rpc.channel.searchVKMTrack([
-                params.hasOwnProperty('vkmRatanumero') ? params.vkmRatanumero : searchValue.ratanumero || '',
-                params.hasOwnProperty('vkmRatakilometri') ? parseInt(params.vkmRatakilometri) : searchValue.ratakilometri || 1,
-                params.hasOwnProperty('vkmRatametri') ? parseInt(params.vkmRatametri) : searchValue.ratametri || 0
-            ], handleResponse, (err) => {
-                setIsSearching(false);
-                if(err){
-                    setError(err);
-                }
-            });
-        } else {
-            setIsSearching(false);
-            setError('required_fields_missing');
-        }
-    };
-
     return (
         <StyledContainer>
             <>
             {
-                error &&
+                vkmTrackError &&
                 <StyledDropdownContentItem>
-                <StyledDropdownContentItemTitle type='noResults'>{error === 'required_fields_missing' ? 'Tarkista tarvittavat kentät' : strings.search.vkm.trackError.text}</StyledDropdownContentItemTitle>
-            </StyledDropdownContentItem>
+                    <StyledDropdownContentItemTitle type='noResults'>{vkmTrackError ? 'Tarkista tarvittavat kentät' : strings.search.vkm.trackError.text}</StyledDropdownContentItemTitle>
+                    <StyledDropdownContentItemSubtitle type='noResults'>{vkmTrackError}</StyledDropdownContentItemSubtitle>
+                </StyledDropdownContentItem>
                 }
                 <StyledDropdownContentItem>
-                    <StyledLabel htmlFor='vkm-track-number'>{strings.search.vkm.ratanumero} *:</StyledLabel>
+                    <StyledLabel
+                        htmlFor='vkm-track-number'
+                        error={vkmTrackError && (searchValue.ratanumero === '' || !searchValue)}
+                    >
+                        {strings.search.vkm.ratanumero} *:
+                    </StyledLabel>
                     <StyledInput
                         id='vkm-track-number'
                         placeholder={strings.search.vkm.ratanumero}
@@ -165,7 +111,7 @@ const VKMTrackSearch = ({
                         value={searchValue.ratanumero || ''}
                         onKeyPress={e => {
                                 if (e.key === 'Enter') {
-                                    handleVKMSearch({vkmRatanumero: e.target.value, vkmRatakilometri: searchValue.ratakilometri, vkmRatametri: searchValue.ratametri});
+                                    handleVKMTrackSearch(searchValue);
                                 }
                             }
                         }
@@ -173,7 +119,12 @@ const VKMTrackSearch = ({
                     />
                 </StyledDropdownContentItem>
                 <StyledDropdownContentItem>
-                    <StyledLabel htmlFor='vkm-track-kilometer'>{strings.search.vkm.ratakilometri} *:</StyledLabel>
+                    <StyledLabel
+                        htmlFor='vkm-track-kilometer'
+                        error={vkmTrackError && (searchValue.ratakilometri === '' || !searchValue)}
+                    >
+                        {strings.search.vkm.ratakilometri} *:
+                    </StyledLabel>
                     <StyledInput
                         id='vkm-track-kilometer'
                         placeholder={strings.search.vkm.ratakilometri}
@@ -194,7 +145,7 @@ const VKMTrackSearch = ({
                         type='number'
                         onKeyPress={e => {
                                 if (e.key === 'Enter') {
-                                    handleVKMSearch({vkmRatanumero: searchValue.ratanumero, vkmRatakilometri: e.target.value, vkmRatametri: searchValue.ratametri});
+                                    handleVKMTrackSearch(searchValue);
                                 }
                             }
                         }
@@ -202,7 +153,12 @@ const VKMTrackSearch = ({
                     />
                 </StyledDropdownContentItem>
                 <StyledDropdownContentItem>
-                    <StyledLabel htmlFor='vkm-track-meter'>{strings.search.vkm.ratametri} *:</StyledLabel>
+                    <StyledLabel
+                        htmlFor='vkm-track-meter'
+                        error={vkmTrackError && (searchValue.ratametri === '' || !searchValue)}
+                    >
+                        {strings.search.vkm.ratametri} *:
+                    </StyledLabel>
                     <StyledInput
                         id='vkm-track-meter'
                         placeholder={strings.search.vkm.ratametri}
@@ -223,7 +179,7 @@ const VKMTrackSearch = ({
                         type='number'
                         onKeyPress={e => {
                                 if (e.key === 'Enter') {
-                                    handleVKMSearch({vkmRatanumero: searchValue.ratanumero, vkmRatakilometri: searchValue.ratakilometri, vkmRatametri: e.target.value});
+                                    handleVKMTrackSearch(searchValue);
                                 }
                             }
                         }